function searchHandler shared to searchPostContent.mjs and search.mjs && JSDoc
