+++ js/utils/search.mjs
@@ -3,17 +3,20 @@ import { renderPost } from "../posts/renderPost.mjs";
 
 
 /**
- * Searches for posts matching the user input and updates the UI accordingly.
+ * Searches for posts matching a given input value and updates the UI.
+ * The function filters posts based on their author, title, body, and tags.
  *
  * @function search
  * @async
- * @param {string} method - The HTTP method "GET".
+ * @param {string} method - The HTTP method to use (e.g., "GET").
  * @param {string} postsUrl - The URL to fetch posts from.
  * @param {string} inputValue - The search query input by the user.
  * @param {HTMLElement} cardContainer - The HTML container element for displaying posts.
  * @param {HTMLElement} allPostsHeader - The HTML element displaying the header for all posts.
- * @param {HTMLElement} filterOption - The HTML element for the filter options.
+ * @param {HTMLElement} filterOption - The HTML element for filter options.
+ * @returns {Promise<void>} - Resolves when the posts have been filtered and rendered.
  * @example
+ * // Example usage within an async function
  * const method = "GET";
  * const postsUrl = "https://api.example.com/posts";
  * const inputValue = "search term";
@@ -22,7 +25,7 @@ import { renderPost } from "../posts/renderPost.mjs";
  * const filterOption = document.querySelector('#filterOption');
  * await search(method, postsUrl, inputValue, cardContainer, allPostsHeader, filterOption);
  */
-async function search(method, postsUrl, inputValue, cardContainer, allPostsHeader, filterOption) {
+export async function search(method, postsUrl, inputValue, cardContainer, allPostsHeader, filterOption) {
   const data = await authWithToken(method, postsUrl);
   const json = data.json;
 
@@ -47,32 +50,4 @@ async function search(method, postsUrl, inputValue, cardContainer, allPostsHeade
   }
 
   renderPost(searchedContent);
-}
-
-
-/**
- * Initializes the search handler and sets up the required event listeners.
- *
- * @param {string} postsUrl - The URL to fetch posts from.
- * @example
- * const postsUrl = "https://api.example.com/posts";
- * searchHandler(postsUrl);
- */
-export function searchHandler(postsUrl) {
-  const searchForm = document.querySelector('#search-form');
-  const searchInput = document.querySelector('input[type="search"]');
-  const filterOption = document.querySelector('#filterOption');
-  const allPostsHeader = document.getElementById('all-posts-header');
-  const cardContainer = document.querySelector('.card-container');
-
-  const method = "GET";
-
-  const handleSearch = async function (e) {
-    e.preventDefault();
-    const inputValue = searchInput.value.toLowerCase();
-    await search(method, postsUrl, inputValue, cardContainer, allPostsHeader, filterOption);
-  };
-
-  searchInput.addEventListener('input', handleSearch);
-  searchForm.addEventListener('submit', handleSearch);
 }
